cacheable - moving back to internal hash on CacheableMemory
